separate out month-based offset tests, parametrize them
